Move stnode.pyi from _typing to its proper place and fix symlink
