+++ src/roman_datamodels/_typing/stnode.py
@@ -1,1 +1,1 @@
-stnode.pyi
+../stnode.pyi
