+++ src/roman_datamodels/stnode.pyi
@@ -1,114 +1,12 @@
-import typing as T
+from __future__ import annotations
 
-if T.TYPE_CHECKING:
-    from astropy.modeling import Model  # type: ignore
-    from astropy.time import Time  # type: ignore
-    from astropy.units import Quantity  # type: ignore
-    from numpy import ndarray  # type: ignore
+import astropy.units as u
+import numpy as np
+from astropy.modeling import Model  # type: ignore
+from astropy.time import Time  # type: ignore
 
-    from roman_datamodels import stnode  # type: ignore
-
-_OptionalQuantity = Quantity | None
-
-class _ndarray_uint8(ndarray):
-    pass
-
-class _ndarray_uint8_3D(_ndarray_uint8):
-    pass
-
-class _ndarray_uint16(ndarray):
-    pass
-
-class _ndarray_uint16_2D(_ndarray_uint16):
-    pass
-
-class _ndarray_uint16_3D(_ndarray_uint16):
-    pass
-
-class _ndarray_uint32(ndarray):
-    pass
-
-class _ndarray_uint32_2D(_ndarray_uint32):
-    pass
-
-class _ndarray_uint32_3D(_ndarray_uint32):
-    pass
-
-class _ndarray_float32(ndarray):
-    pass
-
-class _ndarray_float32_2D(_ndarray_float32):
-    pass
-
-class _ndarray_float32_3D(_ndarray_float32):
-    pass
-
-class _Quantity_float32(Quantity):
-    pass
-
-class _Quantity_float32_electron(_Quantity_float32):
-    pass
-
-class _Quantity_float32_electron_2D(_Quantity_float32_electron):
-    pass
-
-class _Quantity_float32_electron_3D(_Quantity_float32_electron):
-    pass
-
-class _Quantity_float32_electron_DN(_Quantity_float32):
-    pass
-
-class _Quantity_float32_electron_DN_2D(_Quantity_float32_electron_DN):
-    pass
-
-class _Quantity_float32_electron_s(_Quantity_float32):
-    pass
-
-class _Quantity_float32_electron_s_2D(_Quantity_float32_electron_s):
-    pass
-
-class _Quantity_float32_electron_s_3D(_Quantity_float32_electron_s):
-    pass
-
-class _Quantity_float32_electron2_s2(_Quantity_float32):
-    pass
-
-class _Quantity_float32_electron2_s2_2D(_Quantity_float32_electron2_s2):
-    pass
-
-class _Quantity_float32_electron2_s2_3D(_Quantity_float32_electron2_s2):
-    pass
-
-class _Quantity_float32_DN(_Quantity_float32):
-    pass
-
-class _Quantity_float32_DN_2D(_Quantity_float32_DN):
-    pass
-
-class _Quantity_float32_DN_3D(_Quantity_float32_DN):
-    pass
-
-class _Quantity_uint16(Quantity):
-    pass
-
-class _Quantity_uint16_DN(_Quantity_uint16):
-    pass
-
-class _Quantity_uint16_DN_3D(_Quantity_uint16):
-    pass
-
-class _Quantity_uint16_DN_5D(_Quantity_uint16_DN):
-    pass
-
-_GwMode = T.Literal[
-    "WIM-ACQ",
-    "WIM-TRACK",
-    "WSM-ACQ-1",
-    "WSM-ACQ-2",
-    "WSM-TRACK",
-    "DEFOCUSED-MODERATE",
-    "DEFOCUSED-LARGE",
-]
+from roman_datamodels import stnode  # type: ignore
+from roman_datamodels._typing import base, enums
 
 class FileDate(stnode.TaggedScalarNode, Time):
     pass
@@ -122,9 +20,7 @@ class Filename(stnode.TaggedScalarNode, str):
 class ModelType(stnode.TaggedScalarNode, str):
     pass
 
-_Origin = T.Literal["STSCI", "IPAC/SSC"]
-
-class Origin(stnode.TaggedScalarNode, _Origin):
+class Origin(stnode.TaggedScalarNode, str):
     pass
 
 class PrdSoftwareVersion(stnode.TaggedScalarNode, str):
@@ -133,9 +29,7 @@ class PrdSoftwareVersion(stnode.TaggedScalarNode, str):
 class SdfSoftwareVersion(stnode.TaggedScalarNode, str):
     pass
 
-_Telescope = T.Literal["ROMAN"]
-
-class Telescope(stnode.TaggedScalarNode, _Telescope):
+class Telescope(stnode.TaggedScalarNode, str):
     pass
 
 class _Basic(stnode.DNode):
@@ -148,61 +42,31 @@ class _Basic(stnode.DNode):
     sdf_software_version: SdfSoftwareVersion
     telescope: Telescope
 
-_ApertureName = T.Literal[
-    "WFI_01_FULL",
-    "WFI_02_FULL",
-    "WFI_03_FULL",
-    "WFI_04_FULL",
-    "WFI_05_FULL",
-    "WFI_06_FULL",
-    "WFI_07_FULL",
-    "WFI_08_FULL",
-    "WFI_09_FULL",
-    "WFI_10_FULL",
-    "WFI_11_FULL",
-    "WFI_12_FULL",
-    "WFI_13_FULL",
-    "WFI_14_FULL",
-    "WFI_15_FULL",
-    "WFI_16_FULL",
-    "WFI_17_FULL",
-    "WFI_18_FULL",
-    "BORESIGHT",
-    "CGI_CEN",
-    "WFI_CEN",
-]
-
 class Aperture(stnode.TaggedObjectNode):
-    name: _ApertureName
+    name: enums.ApertureName
     position_angle: float
 
-_CalStepVal = T.Literal["N/A", "COMPLETE", "SKIPPED", "INCOMPLETE"]
-
 class CalStep(stnode.TaggedObjectNode):
-    assign_wcs: _CalStepVal
-    flat_field: _CalStepVal
-    dark: _CalStepVal
-    dq_init: _CalStepVal
-    jump: _CalStepVal
-    linearity: _CalStepVal
-    photom: _CalStepVal
-    source_detection: _CalStepVal
-    ramp_fit: _CalStepVal
-    saturation: _CalStepVal
-
-_ReferenceFrame = T.Literal["ICRS"]
+    assign_wcs: enums.CalStep
+    flat_field: enums.CalStep
+    dark: enums.CalStep
+    dq_init: enums.CalStep
+    jump: enums.CalStep
+    linearity: enums.CalStep
+    photom: enums.CalStep
+    source_detection: enums.CalStep
+    ramp_fit: enums.CalStep
+    saturation: enums.CalStep
 
 class Coordinates(stnode.TaggedObjectNode):
-    reference_frame: _ReferenceFrame
-
-_EphemerisType = T.Literal["DEFINITIVE", "PREDICTED"]
+    reference_frame: enums.ReferenceFrame
 
 class Ephemeris(stnode.TaggedObjectNode):
     earth_angle: float
     moon_angle: float
     ephemeris_reference_frame: str
     sun_angle: float
-    type: _EphemerisType
+    type: enums.EphemerisType
     time: float
     spatial_x: float
     spatial_y: float
@@ -211,18 +75,9 @@ class Ephemeris(stnode.TaggedObjectNode):
     velocity_y: float
     velocity_z: float
 
-_ExposureType = T.Literal[
-    "WFI_IMAGE",
-    "WFI_GRISM",
-    "WFI_PRISM",
-    "WFI_DARK",
-    "WFI_FLAT",
-    "WFI_WFSC",
-]
-
 class Exposure(stnode.TaggedObjectNode):
     id: int
-    type: _ExposureType
+    type: enums.ExposureType
     start_time: Time
     mid_time: Time
     end_time: Time
@@ -253,7 +108,7 @@ class Exposure(stnode.TaggedObjectNode):
 
 class Guidestar(stnode.TaggedObjectNode):
     gw_id: str
-    gw_fgs_mode: _GwMode
+    gw_fgs_mode: enums.GuidewindowMode
     gw_science_file_source: str
     gs_id: str
     gs_catalog_version: str
@@ -281,47 +136,10 @@ class Guidestar(stnode.TaggedObjectNode):
     gw_window_xsize: int
     gw_window_ysize: int
 
-_WfiModeName = T.Literal["WFI"]
-_WfiDetector = T.Literal[
-    "WFI01",
-    "WFI02",
-    "WFI03",
-    "WFI04",
-    "WFI05",
-    "WFI06",
-    "WFI07",
-    "WFI08",
-    "WFI09",
-    "WFI10",
-    "WFI11",
-    "WFI12",
-    "WFI13",
-    "WFI14",
-    "WFI15",
-    "WFI16",
-    "WFI17",
-    "WFI18",
-]
-_WfiOpticalElement = T.Literal[
-    "F062",
-    "F087",
-    "F106",
-    "F129",
-    "F146",
-    "F158",
-    "F184",
-    "F213",
-    "GRISM",
-    "PRISM",
-    "DARK",
-]
-
 class WfiMode(stnode.TaggedObjectNode):
-    name: _WfiModeName
-    detector: _WfiDetector
-    optical_element: _WfiOpticalElement
-
-_ObservationSurvey = T.Literal["HLS", "EMS", "SN", "N/A"]
+    name: enums.WfiMode
+    detector: enums.WfiDetector
+    optical_element: enums.WfiOpticalElement
 
 class Observation(stnode.TaggedObjectNode):
     obs_id: str
@@ -338,7 +156,7 @@ class Observation(stnode.TaggedObjectNode):
     exposure: int
     template: str
     observation_label: str
-    survey: _ObservationSurvey
+    survey: enums.ObservationSurvey
 
 class Pointing(stnode.TaggedObjectNode):
     ra_v1: float
@@ -370,13 +188,10 @@ class RefFile(stnode.TaggedObjectNode):
     area: str
     saturation: str
 
-_TargetType = T.Literal["FIXED", "MOVING", "GENERIC"]
-_TargetSourceType = T.Literal["EXTENDED", "POINT", "UNKNOWN"]
-
 class Target(stnode.TaggedObjectNode):
     proposer_name: str
     catalog_name: str
-    type: _TargetType
+    type: enums.TargetType
     ra: float
     dec: float
     ra_uncertainty: float
@@ -386,19 +201,16 @@ class Target(stnode.TaggedObjectNode):
     proper_motion_epoch: str
     proposer_ra: float
     proposer_dec: float
-    source_type: _TargetSourceType
+    source_type: enums.SourceType
 
 class VelocityAberration(stnode.TaggedObjectNode):
     ra_offset: float
     dec_offset: float
     scale_factor: float
 
-_EngineeringQuality = T.Literal["OK", "SUSPECT"]
-_PointingEngdbQuality = T.Literal["CALCULATED", "PLANNED"]
-
 class Visit(stnode.TaggedObjectNode):
-    engineering_quality: _EngineeringQuality
-    pointing_engdb_quality: _PointingEngdbQuality
+    engineering_quality: enums.EngineeringQuality
+    pointing_engdb_quality: enums.PointingEngdbQuality
     type: str
     start_time: Time
     end_time: Time
@@ -445,7 +257,7 @@ class _GuidewindowMeta(_Common):
     signal_resultant_exp_time: float
     gw_acq_number: int
     gw_science_file_source: str
-    gw_mode: _GwMode
+    gw_mode: enums.GuidewindowMode
     gw_window_xstart: int
     gw_window_ystart: int
     gw_window_xstop: int
@@ -455,66 +267,57 @@ class _GuidewindowMeta(_Common):
 
 class Guidewindow(stnode.TaggedObjectNode):
     meta: _GuidewindowMeta
-    pedestal_frames: _Quantity_uint16_DN_5D
-    signal_frames: _Quantity_uint16_DN_5D
-    amp33: _Quantity_uint16_DN_5D
+    pedestal_frames: base.Quantity[np.ndarray[1, np.uint16], u.DN]
+    signal_frames: base.Quantity[np.ndarray[1, np.uint16], u.DN]
+    amp33: base.Quantity[np.ndarray[1, np.uint16], u.DN]
 
 class Ramp(stnode.TaggedObjectNode):
     meta: _Common
-    data: _Quantity_float32_electron_3D
-    pixeldq: _ndarray_uint16_2D
-    groupdq: _ndarray_uint8_3D
-    err: _Quantity_float32_electron_3D
-    amp33: _Quantity_uint16_DN_3D
-    border_ref_pix_left: _Quantity_float32_DN_3D
-    border_ref_pix_right: _Quantity_float32_DN_3D
-    border_ref_pix_top: _Quantity_float32_DN_3D
-    border_ref_pix_bottom: _Quantity_float32_DN_3D
-    dq_border_ref_pix_left: _ndarray_uint16_2D
-    dq_border_ref_pix_right: _ndarray_uint16_2D
-    dq_border_ref_pix_top: _ndarray_uint16_2D
-    dq_border_ref_pix_bottom: _ndarray_uint16_2D
+    data: base.Quantity[np.ndarray[3, np.float32], u.electron]
+    pixeldq: np.ndarray[2, np.uint16]
+    groupdq: np.ndarray[2, np.uint8]
+    err: base.Quantity[np.ndarray[3, np.float32], u.electron]
+    amp33: base.Quantity[np.ndarray[3, np.uint16], u.DN]
+    border_ref_pix_left: base.Quantity[np.ndarray[3, np.float32], u.DN]
+    border_ref_pix_right: base.Quantity[np.ndarray[3, np.float32], u.DN]
+    border_ref_pix_top: base.Quantity[np.ndarray[3, np.float32], u.DN]
+    border_ref_pix_bottom: base.Quantity[np.ndarray[3, np.float32], u.DN]
+    dq_border_ref_pix_left: np.ndarray[2, np.uint16]
+    dq_border_ref_pix_right: np.ndarray[2, np.uint16]
+    dq_border_ref_pix_top: np.ndarray[2, np.uint16]
+    dq_border_ref_pix_left: np.ndarray[2, np.uint16]
 
 class RampFitOutput(stnode.TaggedObjectNode):
     meta: _Common
-    slope: _Quantity_float32_electron_s_3D
-    sigslope: _Quantity_float32_electron_s_3D
-    yint: _Quantity_float32_electron_3D
-    sigyint: _Quantity_float32_electron_3D
-    pedestal: _Quantity_float32_electron_2D
-    weights: _ndarray_float32_3D
-    crmag: _Quantity_float32_electron_3D
-    var_poisson: _Quantity_float32_electron2_s2_3D
-    var_rnoise: _Quantity_float32_electron2_s2_3D
-
-_ResampleWeightType = T.Literal["exptime", "ivm"]
+    slope: base.Quantity[np.ndarray[3, np.float32], u.electron / u.s]
+    sigslope: base.Quantity[np.ndarray[3, np.float32], u.electron / u.s]
+    yint: base.Quantity[np.ndarray[3, np.float32], u.electron]
+    sigyint: base.Quantity[np.ndarray[3, np.float32], u.electron]
+    pedestal: base.Quantity[np.ndarray[2, np.float32], u.electron]
+    weights: np.ndarray[3, np.float32]
+    crmag: base.Quantity[np.ndarray[3, np.float32], u.electron]
+    var_poisson: base.Quantity[np.ndarray[3, np.float32], u.electron**2 / u.s**2]
+    var_rnoise: base.Quantity[np.ndarray[3, np.float32], u.electron**2 / u.s**2]
 
 class Resample(stnode.TaggedObjectNode):
     pixel_scale_ratio: float
     pixfrac: float
     pointings: int
     product_exposure_time: int
-    weight_type: _ResampleWeightType
+    weight_type: enums.ResampleWeightType
 
 class WfiScienceRaw(stnode.TaggedObjectNode):
     meta: _Common
-    data: _Quantity_uint16_DN_3D
-    amp33: _Quantity_uint16_DN_3D
-
-_ConversionMegaJanskys = _OptionalQuantity["MJy"]
-
-_MJy_per_sr = T.Literal["MJy / sr"]
-_uJy_per_arcsec2 = T.Literal["uJy / arcsec2"]
-_sr = T.Literal["sr"]
-_arcsec2 = T.Literal["arcsec2"]
+    data: base.Quantity[np.ndarray[3, np.uint16], u.DN]
+    amp33: base.Quantity[np.ndarray[3, np.uint16], u.DN]
 
 class Photometry(stnode.TaggedObjectNode):
-    conversion_megajanskys: _OptionalQuantity[_MJy_per_sr]
-    conversion_microjanskys: _OptionalQuantity[_uJy_per_arcsec2]
-    pixelarea_steradians: _OptionalQuantity[_sr]
-    pixelarea_arcsecsq: _OptionalQuantity[_arcsec2]
-    conversion_megajanskys_uncertainty: _OptionalQuantity[_MJy_per_sr]
-    conversion_microjanskys_uncertainty: _OptionalQuantity[_uJy_per_arcsec2]
+    conversion_megajanskys: u.Quantity[u.MJy / u.sr] | None
+    conversion_microjanskys: u.Quantity[u.uJy / u.arcsec**2] | None
+    pixelarea_steradians: u.Quantity[u.sr] | None
+    pixelarea_arcsecsq: u.Quantity[u.arcsec**2] | None
+    conversion_megajanskys_uncertainty: u.Quantity[u.MJy / u.sr] | None
+    conversion_microjanskys_uncertainty: u.Quantity[u.uJy / u.arcsec**2] | None
 
 class SourceDetection(stnode.TaggedObjectNode):
     tweakreg_catalog_name: str
@@ -528,21 +331,21 @@ class CalLogs(stnode.TaggedListNode):
 
 class WfiImage(stnode.TaggedObjectNode):
     meta: _WfiImageMeta
-    data: _Quantity_float32_electron_s_2D
-    dq: _ndarray_uint16_2D
-    err: _Quantity_float32_electron_s_2D
-    var_poisson: _Quantity_float32_electron2_s2_2D
-    var_rnoise: _Quantity_float32_electron2_s2_2D
-    var_flat: _Quantity_float32_electron2_s2_2D
-    amp33: _Quantity_uint16_DN_3D
-    border_ref_pix_left: _Quantity_float32_DN_3D
-    border_ref_pix_right: _Quantity_float32_DN_3D
-    border_ref_pix_top: _Quantity_float32_DN_3D
-    border_ref_pix_bottom: _Quantity_float32_DN_3D
-    dq_border_ref_pix_left: _ndarray_uint32_2D
-    dq_border_ref_pix_right: _ndarray_uint32_2D
-    dq_border_ref_pix_top: _ndarray_uint32_2D
-    dq_border_ref_pix_bottom: _ndarray_uint32_2D
+    data: base.Quantity[np.ndarray[2, np.float32], u.electron / u.s]
+    dq: np.ndarray[2, np.uint16]
+    err: base.Quantity[np.ndarray[2, np.float32], u.electron / u.s]
+    var_poisson: base.Quantity[np.ndarray[2, np.float32], u.electron**2 / u.s**2]
+    var_rnoise: base.Quantity[np.ndarray[2, np.float32], u.electron**2 / u.s**2]
+    var_flat: base.Quantity[np.ndarray[2, np.float32], u.electron**2 / u.s**2]
+    amp33: base.Quantity[np.ndarray[3, np.uint16], u.DN]
+    border_ref_pix_left: base.Quantity[np.ndarray[3, np.float32], u.DN]
+    border_ref_pix_right: base.Quantity[np.ndarray[3, np.float32], u.DN]
+    border_ref_pix_top: base.Quantity[np.ndarray[3, np.float32], u.DN]
+    border_ref_pix_bottom: base.Quantity[np.ndarray[3, np.float32], u.DN]
+    dq_border_ref_pix_left: np.ndarray[3, np.uint32]
+    dq_border_ref_pix_right: np.ndarray[3, np.uint32]
+    dq_border_ref_pix_top: np.ndarray[3, np.uint32]
+    dq_border_ref_pix_bottom: np.ndarray[3, np.uint32]
     cal_logs: CalLogs
 
 class _WfiMosiacMeta(_Common):
@@ -551,27 +354,18 @@ class _WfiMosiacMeta(_Common):
 
 class WfiMosaic(stnode.TaggedObjectNode):
     meta: _WfiMosiacMeta
-    data: _Quantity_float32_electron_s_2D
-    err: _Quantity_float32_electron_s_2D
-    context: _ndarray_uint32_3D
-    weight: _ndarray_float32_2D
-    var_poisson: _Quantity_float32_electron2_s2_2D
-    var_rnoise: _Quantity_float32_electron2_s2_2D
-    var_flat: _Quantity_float32_electron2_s2_2D
+    data: base.Quantity[np.ndarray[2, np.float32], u.electron / u.s]
+    err: base.Quantity[np.ndarray[2, np.float32], u.electron / u.s]
+    context: np.ndarray[3, np.uint32]
+    weight: np.ndarray[2, np.float32]
+    var_poisson: base.Quantity[np.ndarray[2, np.float32], u.electron**2 / u.s**2]
+    var_rnoise: base.Quantity[np.ndarray[2, np.float32], u.electron**2 / u.s**2]
+    var_flat: base.Quantity[np.ndarray[2, np.float32], u.electron**2 / u.s**2]
     cal_logs: CalLogs
 
-_p_exptype = T.Literal[
-    "WFI_IMAGE",
-    "WFI_GRISM",
-    "WFI_PRISM",
-    "WFI_DARK",
-    "WFI_FLAT",
-    "WFI_WFSC",
-]
-
 class _RefExposureType(stnode.DNode):
-    type: _ExposureType
-    p_exptype: _p_exptype
+    type: enums.ExposureType
+    p_exptype: enums.ExposureType
 
 class _DarkExposure(_RefExposureType):
     ngroups: int
@@ -581,169 +375,136 @@ class _DarkExposure(_RefExposureType):
     ma_table_number: int
 
 class _Instrument(stnode.DNode):
-    name: _WfiModeName
-    detector: _WfiDetector
-    optical_element: _WfiOpticalElement
-
-_Pedigree = T.Literal["GROUND", "MODEL", "DUMMY", "SIMULATION"]
+    name: enums.WfiMode
+    detector: enums.WfiDetector
+    optical_element: enums.WfiOpticalElement
 
 class _RefCommon(stnode.DNode):
-    pedigree: _Pedigree
+    pedigree: enums.Pedigree
     description: str
     author: str
     useafter: Time
-    telescope: _Telescope
+    telescope: enums.Telescope
     origin: str
     instrument: _Instrument
 
-_DarkRefType = T.Literal["DARK"]
-
 class _DarkRefMeta(_RefCommon):
-    reftype: _DarkRefType
+    reftype: enums.DarkRefType
     exposure: _DarkExposure
     instrument: _Instrument
 
 class DarkRef(stnode.TaggedObjectNode):
     meta: _DarkRefMeta
-    data: _Quantity_float32_DN_3D
-    dq: _ndarray_uint32_2D
-    err: _Quantity_float32_DN_3D
-
-_DistortionInput = T.Literal["pix"]
-_DistortionOutput = T.Literal["arcsec"]
-
-_DistortionRefType = T.Literal["DISTORTION"]
+    data: base.Quantity[np.ndarray[3, np.float32], u.DN]
+    dq: base.Quantity[np.ndarray[2, np.uint32], u.DN]
+    err: base.Quantity[np.ndarray[3, np.float32], u.DN]
 
 class _DistortionRefMeta(_RefCommon):
-    reftype: _DistortionRefType
-    input_units: _DistortionInput
-    output_units: _DistortionOutput
+    reftype: enums.DistortionRefType
+    input_units: base.Unit[u.pix]
+    output_units: base.Unit[u.arcsec]
 
 class DistortionRef(stnode.TaggedObjectNode):
     meta: _DistortionRefMeta
     coordinate_distortion_transform: Model
 
-_FlatRefType = T.Literal["FLAT"]
-
 class _FlatRefMeta(_RefCommon):
-    reftype: _FlatRefType
+    reftype: enums.FlatRefType
     instrument: _Instrument
 
 class FlatRef(stnode.TaggedObjectNode):
     meta: _FlatRefMeta
-    data: _ndarray_float32_2D
-    dq: _ndarray_uint32_2D
-    err: _ndarray_float32_2D
-
-_GainRefType = T.Literal["GAIN"]
+    data: np.ndarray[2, np.float32]
+    dq: np.ndarray[2, np.uint32]
+    err: np.ndarray[2, np.float32]
 
 class _GainRefMeta(_RefCommon):
-    reftype: _GainRefType
+    reftype: enums.GainRefType
 
 class GainRef(stnode.TaggedObjectNode):
     meta: _GainRefMeta
-    data: _Quantity_float32_electron_DN_2D
-
-_DNunit = T.Literal["DN"]
-
-_InverseLinearityRefType = T.Literal["INVERSE_LINEARITY"]
+    data: base.Quantity[np.ndarray[2, np.float32], u.electron / u.DN]
 
 class _InverseLinearityRefMeta(_RefCommon):
-    reftype: _InverseLinearityRefType
-    input_units: _DNunit
-    output_units: _DNunit
+    reftype: enums.InverseLinearityRefType
+    input_units: base.Unit[u.DN]
+    output_units: base.Unit[u.DN]
 
 class InverseLinearityRef(stnode.TaggedObjectNode):
     meta: _InverseLinearityRefMeta
-    coeffs: _ndarray_float32_3D
-    dq: _ndarray_uint32_2D
-
-_IpcRefType = T.Literal["IPC"]
+    coeffs: np.ndarray[3, np.float32]
+    dq: np.ndarray[2, np.uint32]
 
 class _IpcRefMeta(_RefCommon):
-    reftype: _IpcRefType
+    reftype: enums.IPCRefType
     instrument: _Instrument
 
 class IpcRef(stnode.TaggedObjectNode):
     meta: _IpcRefMeta
-    data: _ndarray_float32_2D
-
-_LinearityRefType = T.Literal["LINEARITY"]
+    data: np.ndarray[3, np.float32]
 
 class _LinearityRefMeta(_RefCommon):
-    reftype: _LinearityRefType
-    input_units: _DNunit
-    output_units: _DNunit
+    reftype: enums.LinearityRefType
+    input_units: base.Unit[u.DN]
+    output_units: base.Unit[u.DN]
 
 class LinearityRef(stnode.TaggedObjectNode):
     meta: _LinearityRefMeta
-    coeffs: _ndarray_float32_3D
-    dq: _ndarray_uint32_2D
-
-_MaskRefType = T.Literal["MASK"]
+    coeffs: np.ndarray[3, np.float32]
+    dq: np.ndarray[2, np.uint32]
 
 class _MaskRefMeta(_RefCommon):
-    reftype: _MaskRefType
+    reftype: enums.MaskRefType
 
 class MaskRef(stnode.TaggedObjectNode):
     meta: _MaskRefMeta
-    dq: _ndarray_uint32_2D
+    dq: np.ndarray[2, np.uint32]
 
 class _PixelareaRefPhotometry(stnode.DNode):
-    pixelarea_steradians: _OptionalQuantity[_sr]
-    pixelarea_arcsecsq: _OptionalQuantity[_arcsec2]
-
-_PixelareaRefType = T.Literal["AREA"]
+    pixelarea_steradians: u.Quantity[u.sr] | None
+    pixelarea_arcsecsq: u.Quantity[u.arcsec**2] | None
 
 class _PixelareaRefMeta(_RefCommon):
-    reftype: _PixelareaRefType
+    reftype: enums.PixelareaRefType
     photometry: _PixelareaRefPhotometry
     instrument: _Instrument
 
 class PixelareaRef(stnode.TaggedObjectNode):
     meta: _PixelareaRefMeta
-    data: _ndarray_float32_2D
-
-_ReadnoiseRefType = T.Literal["READNOISE"]
+    data: np.ndarray[2, np.float32]
 
 class _ReadnoiseRefMeta(_RefCommon):
-    reftype: _ReadnoiseRefType
+    reftype: enums.ReadnoiseRefType
     exposure: _RefExposureType
 
 class ReadnoiseRef(stnode.TaggedObjectNode):
     meta: _ReadnoiseRefMeta
-    data: _Quantity_float32_DN_2D
-
-_SaturationRefType = T.Literal["SATURATION"]
+    data: base.Quantity[np.ndarray[2, np.float32], u.DN]
 
 class _SaturationRefMeta(_RefCommon):
-    reftype: _SaturationRefType
+    reftype: enums.SaturationRefType
 
 class SaturationRef(stnode.TaggedObjectNode):
     meta: _SaturationRefMeta
-    data: _Quantity_float32_DN_2D
-    dq: _ndarray_uint32_2D
-
-_SuperbiasRefType = T.Literal["BIAS"]
+    data: base.Quantity[np.ndarray[2, np.float32], u.DN]
+    dq: np.ndarray[2, np.uint32]
 
 class _SuperbiasRefMeta(_RefCommon):
-    reftype: _SuperbiasRefType
+    reftype: enums.SuperbiasRefType
 
 class SuperbiasRef(stnode.TaggedObjectNode):
     meta: _SuperbiasRefMeta
-    data: _ndarray_float32_2D
-    dq: _ndarray_uint32_2D
-    err: _ndarray_float32_2D
-
-_WfiImgPhotomRefType = T.Literal["PHOTOM"]
+    data: np.ndarray[2, np.float32]
+    dq: np.ndarray[2, np.uint32]
+    err: np.ndarray[2, np.float32]
 
 class _WfiImgPhotomRefMeta(_RefCommon):
-    reftype: _WfiImgPhotomRefType
+    reftype: enums.WfiPhotomRefType
 
 class _PhotTableEntry(stnode.DNode):
-    photmjsr: _OptionalQuantity[_MJy_per_sr]
-    uncertainty: _OptionalQuantity[_MJy_per_sr]
-    pixelareasr: _OptionalQuantity[_sr]
+    photmjsr: u.Quantity[u.MJy / u.sr] | None
+    uncertainty: u.Quantity[u.MJy / u.sr] | None
+    pixelareasr: u.Quantity[u.sr] | None
 
 class _PhotTable(stnode.DNode):
     F062: _PhotTableEntry
